+++ src/main/java/com/selina/lending/api/controller/CreditCommitmentsController.java
@@ -18,7 +18,7 @@
 package com.selina.lending.api.controller;
 
 import com.selina.lending.internal.dto.ApplicationResponse;
-import com.selina.lending.internal.dto.creaditCommitments.UpdateCreditCommitmentsRequest;
+import com.selina.lending.internal.dto.creditCommitments.UpdateCreditCommitmentsRequest;
 import com.selina.lending.internal.mapper.ApplicationResponseMapper;
 import com.selina.lending.internal.service.creaditCommitments.UpdateCreditCommitmentsService;
 import lombok.extern.slf4j.Slf4j;
