+++ src/main/java/com/selina/lending/internal/api/MiddlewareApplicationServiceApi.java
@@ -34,7 +34,6 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
         url = "${middleware.application.service.url}")
 public interface MiddlewareApplicationServiceApi {
 
-    // TODO: does it realy return body similar to ApplicationResponse
     @PostMapping(path = "/application/runDecisioning/{applicationId}", produces = APPLICATION_JSON_VALUE)
     ApplicationResponse runDecisioningByAppId(@PathVariable("applicationId") String applicationId);
