+++ src/main/java/com/selina/lending/internal/dto/creditCommitments/ApplicantCreditCommitments.java
@@ -15,18 +15,21 @@
  *
  */
 
-package com.selina.lending.internal.dto.creaditCommitments;
+package com.selina.lending.internal.dto.creditCommitments;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
+import com.selina.lending.internal.dto.CreditCommitmentsDto;
+import lombok.Builder;
+import lombok.Value;
 
-import java.util.List;
-
-@NoArgsConstructor
-@SuperBuilder
-@Getter
-public class CreditCommitmentResponse {
-    String id;
-    List<ApplicantCreditCommitments> applicants;
+@Value
+@Builder
+public class ApplicantCreditCommitments {
+    Long id;
+    Boolean primaryApplicant;
+    String experianId;
+    String searchId;
+    String creditScore;
+    String message;
+    String creditCheck;
+    CreditCommitmentsDto creditCommitments;
 }
