+++ src/main/java/com/selina/lending/internal/service/creaditCommitments/UpdateCreditCommitmentsService.java
@@ -17,7 +17,7 @@
 
 package com.selina.lending.internal.service.creaditCommitments;
 
-import com.selina.lending.internal.dto.creaditCommitments.UpdateCreditCommitmentsRequest;
+import com.selina.lending.internal.dto.creditCommitments.UpdateCreditCommitmentsRequest;
 import com.selina.lending.internal.service.application.domain.ApplicationResponse;
 
 public interface UpdateCreditCommitmentsService {
