+++ src/test/java/com/selina/lending/api/controller/CreditCommitmentsControllerUnitTest.java
@@ -17,11 +17,9 @@
 
 package com.selina.lending.api.controller;
 
-import com.selina.lending.internal.dto.creaditCommitments.UpdateCreditCommitmentsRequest;
+import com.selina.lending.internal.dto.creditCommitments.UpdateCreditCommitmentsRequest;
 import com.selina.lending.internal.mapper.MapperBase;
-import com.selina.lending.internal.service.application.domain.ApplicationResponse;
 import com.selina.lending.internal.service.creaditCommitments.UpdateCreditCommitmentsService;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
